Drop support of Jetty NPN
Motivation:

The NPN feature is currently supported in Netty using the Jetty NPN library. This library has been archived and the maintainers are not keen on continuing to support it.

NPN has been deprecated in favour of ALPN and browsers should be using ALPN instead (ALPN is almost 10 years old), the Jetty project has dropped NPN in Jetty 9.3 (2015 - https://webtide.com/phasing-out-spdy-support/) and has not seen evidence of the lack of NPN being an issue.

Changes:

Remove dependency and handling of Jetty NPN.

Results:

Netty does not depend anymore on `org.eclipse.jetty.npn:npn-api` .

Fixes #14237