Slice token before pass it to QuicTokenHandler (#743)
Motivation:

As the QuicTokenHandler implementation might modify the readerIndex of
the token we need to ensure we slice it before pass it to to the
handler. Otherwise we will run into issues that might fail the
connection establishment

Modifications:

- Slice token before pass it to the QuicTokenHandler
- Add unit test

Result:

No more connection issues due of token readerIndex adjustments. Fixes
https://github.com/netty/netty-incubator-codec-quic/issues/742