Adding the Modular Netty guide aimed at users and developers. (#14274)
Motivation:

Users and developers of Netty need guidance when using Netty with JPMS.

Changes:

Provide a guide that explains how to use Netty with JPMS as well as the
constraints, as well how to contribute to Netty. Provide a basic help
for the HTTP server example.

Result:

There is a guide solving this problem.