Remove overrides in AdaptiveByteBuf that caused double index checks
Motivation:

The overrides that we had were kind of wasteful as it would cause to first check the index within the AdaptiveByteBuf and then also in the rootParent ByteBuf. Let's just remove the overrides as AbstractByteBuf (which AdaptiveByteBuf extends) already does the index checking itself and then delegate to the right _* method. If the index check in AdaptiveByteBuf succeed it should never fail in the rootParent, if it does its a bug.

Modifications:

Reduce overhead in AdapativeByteBuf.

Result:

Less overhead