Only add Magazine to Set if we can ensure its removed again
Motivation:

We need to ensure we only add the Magazine to the Set if we are sure it
is removed again once the Thread exit. Otherwise we will end up with a
leak as the Magazine will only be able to collected once the whole
allocator is GC'ed

Modifications:

- Only add Magazine if we are sure onRemoval(...) is called

Result:

No more leak when used from non FastThreadLocalThread.