Make arenas reuse their last chunk more aggressively (#14291)
Motivation:
Particular allocation patterns can cause the PooledByteBufAllocator to
continuously allocate and deallocate chunks, leading to high memory
churn.

Modification:
When the HeapArena destroys a chunk, it does so while keeping one chunk
alive for later reuse. If a chunk is already stowed for reuse, then heap
chunks get destroyed like before. When allocating a new chunk, we
atomically grab any stowed chunks and set the cache field to null. We
assert that the parameters are all the same, which should be the case
since they're all derived from the SizeClass.

We only do this caching for the heap chunks, because the direct chunks
have explicit lifetimes, and it would be too complicated to do this
trick for them. This also respects the opposing desire to release native
memory back to the OS with expediency, when it's no longer needed.

Result:
We add a single-chunk cache that acts as a damper for any chunk
allocation/deallocation oscillations that might occur with certain
allocation patterns.

This fixes #14257