Fix sizeBucket bug in AdaptivePoolingAllocator
Motivation:
The sizeBucket() computation was mistakenly masking the normalized size, rather than the bit-count.
This made the allocator choose very wrong size buckets for larger buffers.
The incorrect size buckets would skew the collected statistics toward smaller allocations, and fool the allocator into optimizing for smaller chunk sizes.

Modification:
The size-bucket masking is moved to the size-bucket number computed from the bits.
This ensures the size-bucket number does not exceed the size-bucket arrays and histograms we allocate.
Specifically, the masking was in a place where it would penalize buffers bigger than 64 KiB, and compute incorrect (too small) size buckets for them.

Additionally, the INIT_DATUM_TARGET has been reduced dramatically.
The datum target decides when the histogram is rotated and the preferred chunk size computed.
Chunks are initially sized to fit 10 buffers sized like the first buffer allocated for a given magazine.
If this holds true, then it's desirable to compute the preferred chunk size before the first chunk is deallocated and the next one allocated.

This change makes the adaptive allocator update its histograms very aggressively in the beginning, until things settle down.

Result:
The AdaptivePoolingAllocator now computes much better chunk sizes for allocation patterns that prefer their buffers bigger than 64 KiB.

This is a back-port of https://github.com/netty/netty/pull/14305,
which in turn is a forward-port and refinement of some of the changes in https://github.com/netty/netty/pull/14291