Fix SocketSslClientRenegotiateTest
Motivation:
The endpoint verification algorithm was rejecting our self-signed certificate for not having any hostname.

Modification:
Disable the endpoint verification because it's not relevant to the test, and it's might be more trouble than it's worth to portably add a hostname for localhost in the cert.

Result:
SocketSslClientRenegotiateTest pass again in builds that use OpenSSL, instead of BoringSSL.