AdaptiveByteBufAllocator: Use Pooled*ByteBuf for retained slices / duplicates if we use a Recycler
Motivation:

We can use Pooled*ByteBuf for retained slices / duplicates when we use the Recycler. This allows us to reduce overhead an GC pressure (the same way as in PooledByteBufAllocator).

Modifications:

Check if we have a Handle and if so use Pooled'ByteVuf for slice / duplicates that are retained

Result:

Less GC pressure