Assert that nameservers are resolved before using these as keys in the map
Motivation:

We should assert that nameserver are resolved before we use these as keys in our map internally in our DnsNameResolver. This is as otherwise hashCode() and equals() will change and so mess up the map.

Modifications:

Add some asserts to ensure we never let things slip through in the future

Result:

More robust code