Magazines must be freed under the expand lock (#14336)
Motivation:
It's unlikely, but possible, for our freeing of magazines to run
concurrently with magazine expansion. If that happens, we might not free
all magazines.

Modification:
Free the magazines under the expansion lock, and short-circuit the
magazine expansion if the allocator has been freed.

Result:
Fixes a data race that could leave magazines and their chunks un-freed.