WritabilityChanged event is triggered with wrong capacity (#748)
Motivation:

There was a regression in `fdb1a7` (00.63.Final) where `writeQueued()`
may call `updateWritabilityIfNeeded(true)` at the end of the loop, even
when no write operations were completed. As a result, the channel
capacity is zero. That has an impact on handlers that intercept the
WritabilityChanged events and call `Channel.bytesBeforeUnwritable()`.

Modifications:

Only call `updateWritabilityIfNeeded(true)` if `written` is true.

Result:

Handlers that intercept WritabilityChanged events see the correct
capacity.