Fix OpenSslClientSessionCache remove
Motivation:
Loop was adding `nativeSslSession` to the `toBeRemoved` list, instead of the loop variable `sslSession`.

Modification:
Add the loop variable instead, because the `nativeSslSession` can very likely be null at that point, and is specifically also not the session we want to remove.

Result:
Fewer bugs in SSL session caching