Fix flaky BootstrapTest
Motivation:
Calls to Channel.isOpen() should be gated by the channel close future, at least if we expected the channel to be closed.

Modification:
Add some awaits on the channel close future, before asserting that a channel is closed.

Result:
The tests should no longer occasionally fail on that assertion.