Refactor DnsNameResolver to be able to use different strategies when … (#14374)
…it comes to creating Channels for queries.

Motivation:

At the moment we reuse the same Channel (and so underyling socket) for
all queries. It has show that this can cause problems in some
deployments, we should allow to use different strategies.

Modifications:

- Refactor code to make it possible to use different strategies when it
comes to creating / reusing channels for resolve operations

Result:

More flexible configuration and possible fix for
https://github.com/netty/netty/issues/14364

---------

Co-authored-by: Idel Pivnitskiy <idel.pivnitskiy@apple.com>