Create new Chunk which shares the internal storage of the previous Chunk when queue for reuse
Motivation:

We need to ensure we never end up with a race / inconsistency related to the reference count when a Chunk is queued to be reused. Previous to this commit we never correctly set the reference count when we could not queue it.

Modifications:

Create a new Chunk and so ensure we never end up with inconsistent state / races in terms of reference count

Result:

More correct code