Fix incorrect cast in NioDomainSocketChannel.parent() (#14409)
Motivation:

NioDomainSocketChannel.parent() returned a ServerSocketChannel, which is
not implemented by NioServerDomainSocketChannel. This makes the parent
method unusable.

Modification:

Change the return type and cast to ServerChannel. Unfortunately this is
a binary incompatible change, but there is no way to make this change in
a binary compatible way, and the method never worked in the first place.

Result:

parent() becomes usable.