Verify default named groups before using them with native SSL implementation
Motivation:

We should verify that the default named groups are actually supported by our native SSL implementation. This might not always be the case as for example when FIPS is used.

Modifications:

- Verify that default named groups are supported
- Fail creation of ReferenceCountedOpenSslContext if setting of groups fails and also include details about why it failed if possible

Result:

Easier to debug miss-configuration of groups and make things work out of the box even if FIPS is used. Related to https://github.com/netty/netty-tcnative/issues/883