Include details on why it was not possible to configure accepted issuers in the SSLException
Motivation:

At the moment it's hard to understand why the configuration of the SSLContext fails when we can not configure the accepted issuers. Let's include more details if possible to make debugging easier.

Modifications:

Check if we have more details and if so include it in the exception message

Result:

Related to https://github.com/netty/netty-tcnative/issues/883