Merge commit from fork
Motivation:

We didnt verify the amount of data buffered while reading files / streams during startup which could allow an attacker to cause an OMME if the attacker was able to create a massive file.

Modifications:

Decorate InputStream with own that will throw an exception once we reach the limit of data read.

Result:

No more risk of OOME