Pass op code through completion callbacks
Motivation:

We call the same completion callbacks for various different op codes. As we might want to do different things depending on the operation that was completes in the future we should better just pass the op code as well.

Modifications:

Pass op code to completion methods

Result:

More flexible in the future