PcapWriteHandler no longer ignores writePcapGlobalHeader
Motivation:

PcapWriteHandler.Builder has a boolean for writePcapGlobalHeader so that the handler can support functionality like appending to an existing PCAP file. This boolean is currently not being checked when writing the global header, preventing appending from working.

Modifications:

PcapWriter checks PcapWriteHandler.writePcapGlobalHeader() returned true.
Added unit test cases for when writePcapGlobalHeader is set to false.

Result:

PcapWriteHandler only writes the global PCAP header on initialization if writePcapGlobalHeader is true and if sharedOutputStream is false.