IoUring: data passed to callbacks should be of type short
Motiviation:

IoUringOps allows to add some data to each op that is passed as part of the completion event. The type of this data is short but we used int for the callbacks.

Modification:

Use short for data in callbacks

Result:

Consistent usage of data type for user supplied data