Added decoder for simplifying decoding DatagramPacket into AddressedEnvelope with a generic MessageToMessageDecoder<ByteBuf>
Motivation:

There is an encoder for encoding an AddressedEnvelope into a DatagramPacket and a decoder for decoding the contents of a DatagramPacket into an Object. There is currently no good mechanism for decoding a DatagramPacket into an AddressedEnvelope. You would ideally be able to wrap a decoder like ProtobufDecoder with a decoder that wraps the protobuf message in an envelope. This is difficult to do outside Netty due to the protection level of decode().

Modifications:

Added decoder for AddressedEnvelope

Result:

Users of netty will be able to more easily propagate datagram addresses through the inbound pipeline.