IoUring: Correctly handle cancelletion of close and connect
Motivation:

When a close op is cancelled we need to not handle it the same as when it failed / worked. Also when a connect is cancelled we need to ensure we not cancel the connect timeout ans also not null out the connect promise.

Modification:

Correct handle cancellation.

Result:

Handle cancellation of ops correctly