IoUring: IoUringRegistration.cancel() didn't remove registration when called only once
Motivation:

There was a bug in the cancel() method that did let it return early by mistake on the first call while it should have only returned early after the first call. This resulted into registrations to not be removed either completely or not in a timely fashion.

Modifications:

Fix early return.

Result:

Registrations are removed correctly