Deprecate ChannelPipeline.add*(EventExecutorGroup....) methods
Motivation:
While it seemed like a great idea to allow to run different handlers with different executors it became clear that this introduce more problems then it solves. If someone wants to run things outside of the EventLoop its much better to dispatch it manually.

Modifications:

Add @deprecated to the methods

Result:

Prepare users for future removal and discurage usage