Implement 'inEventLoop' for UnorderedThreadPoolEventExecutor and deprecate it
Motivation:
The `UnorderedThreadPoolEventExecutor` did not implement the `inEventLoop` method, and instead always returned `false`.
This can have deleterious consequences to many parts of the Netty code that have optimized code paths for event loop threads.

The `UnorderedThreadPoolEventExecutor` also behaves in a way that deviates from normal Netty event loops in important ways, and it has very little use.
Therefor it should be deprecated and removed in Netty 5.

Modification:
Add executor thread-accounting to the `UnorderedThreadPoolEventExecutor` and use the set of event loop threads to implement `inEventLoop` properly.

Result:
The `UnorderedThreadPoolEventExecutor.inEventLoop` method now correctly reports if the current thread is managed by the executor or not.