Add a module for generating X.509 certificates for testing
Motivation:
You need an X.509 certificate in order to test much of anything TLS related.
This is true for both Netty, and anyone else using Netty.
The `SelfSignedCertificate` was fine (if a bit clunky) for just Netty usage, but is reaching the limits of its API when others are using it.
It relies on internal JDK classes (which have since been restricted), or it relies on BouncyCastle being available (as an optional dependency), or it relies on Keytool (which involves shelling out, and cannot be done in-memory).
The `SelfSignedCertificate` is also limited in that it can only produce self-signed certificates, when most certificates used in practice are signed by a chain of issuers.

Modification:
Add a separate module for creating test certificates, and implement the certificate signing from scratch such that only BouncyCastle dependencies are required.
The `CertificateBuilder` permit far more customization of the produced certificates, such as key usages, and other extensions.
It returns `X509Bundle` objects, which contain both the certificate and its corresponding key pair.
These can be used as signatories for other certificates, allowing us to simulate a Certificate Authority.
Support for revocations through CRLs is provided through a `RevocationServer`, using the JDK-built-in HTTP server to avoid circular dependencies.

The `SelfSignedCertificate` and supporting classes are removed, and replaced by pkitesting.

Result:
Netty and its users can now reliably create arbitrary test certificates.