Fix bugs in BoundedInputStream (#14479) (#14480)
Motivation:

A BoundedInputStream was recently introduced to protect against attacks
where a large file was placed on the server. Unfortunately, this has an
issue preventing it from reading the last bytes of a file where the
buffer size would take the total bytes over the set bound, when reading
just the remaining bytes in the file fits underneath the set bound.

Modification:

Adjusted the behaviour of BoundedInputStream to, where appropriate, read
1 more byte than the set bound, to see if the file terminates at this
point, in which case there is no need to throw the exception

Result:

Fixes #14479.