IoUring: Expose features that can be used
Motiviation:

When setup io_uring the kernel does let us know which features can be used. We should expose this so we can make use of it later to detect what we can use.

Modifications:

Change setup method to also return the features.

Result:

Be able to detect what features the kernel supports