Add a RedHat ubi9 based build
Motivation:
CentOS 6 is ancient, and the repo mirrors are even starting to disappear.
Netty 5 should be based on a much more modern OS environment, so let's explore some alternatives like RedHat ubi9.

Modification:
Add a RedHat ubi9 based PR build.
We only build with the BoringSSL variant of tcnative on ubi9, because tcnative wishes to link to openssl 1.0, and that's not available anymore.

Result:
We get to see if this works, and if it does, we can move the other builds over to this base image.