Remove the m2e plug-in work around (#14599) (#14601)
Motivation:
This section is highlight as erroneous and invalid POM by Intellij. It's
been over 12 years since it was added, so surely the work-around is no
longer needed?

Modification:
Remove the explicit call-out to the m2eclipse plugin. Also added empty
definitions for one property that were referenced but not defined, and
as such were also highlighted as errors.

Result:
Cleaner pom.xml file.