Use Arrays in `IpSubnetFilter` (#14593)
Motivation:
`IpSubnetFilter` always initializes two Lists, one for IPv4 and one for
IPv6. This is fine if we have both types of rules. However, if we only
process one or the other, one allocation is useless and stays forever
since it's a `final` variable. Likewise, we should use Arrays as it has
less overhead compared to List.

Modification:
Migrate to Arrays and only initialize when required

Result:
Optimize memory allocation