Fix classloader leaks in GlobalEventExecutor
Motivation:
If Netty is on a main classpath in a system that creates ephemeral classloaders, e.g. for plug-ins, then we should avoid holding on to those ephemeral classloaders.

Modification:
Clear the stack trace in the GlobalEventExecutor termination future, because the stack trace would hold on to class which in turn hold on to class loaders, and it would hold on to an ephemeral classloader if the GEE was first initialized by code in an ephemeral classloader.

Also avoid holding on to context classloaders through the Thread.inheritedAccessControlContext when starting the GEE thread.

Result:
Classloader leaks should be fix.

Fixes #14615