OpenSslSession: Add support to defensively check for peer certs
Motivation
----------
In a use-case where an application wants to check if a peer certificate
has been provided without throwing an exception if none are found
(as is currently the behavior when calling SSLSession#getPeerCertificates)
there is currently no API available to do that.

The use-case which prompted this addition is to defensively check if
a mTLS connection has been established with access to the SSLSession.

Modifications
-------------
This changeset introduces a new API to the OpenSslSession which allows
to check if peer certs are available (hasPeerCertificates) and returns
true if this is the case.

Result
------
It is now possible to check if mTLS is enabled (through checking if
peer certs are presented) without throwing an exception if not.