Reentrant close in EmbeddedChannel Motivation:
This is another case like #13730 where a close operation in channelRead can lead to a reentrant call to handlerRemoved. In this case, it's a direct Channel#close call that bypasses the reentrancy guard introduced in #13730, and calls the close logic immediately.

ChannelHandlerContext#close is not affected, but Channel#close is called e.g. by ChannelFutureListener#CLOSE so using the context is not always viable.

Discovered by fuzzing.

Modification:

Instead of cancelling scheduled tasks immediately, make the cancellation part of the maybeRunPendingTasks logic.

Result:

Reentrancy is controlled and the test passes.