Clarify documentation on max HTTP header size (#14670)
Motivation:
Limiting the size of buffers in the HTTP decoders is an important
security consideration. The API documentation on the max header size
could be more clear, so people can make correct decisions about how to
configure it.

Modification:
Updated the API documentation on the HttpDecoderConfig methods, and
HttpRequest/ResponseDecoder constructor parameters, to clarify that the
limit applies to the sum total size of all the headers in an HTTP
message.

Result:
Clearer documentation.

---------

Co-authored-by: Norman Maurer <norman_maurer@apple.com>