IoUring: try to enable IORING_SETUP_SUBMIT_ALL when supported by the running system
Motivation

When porting io_uring to the main branch we introduced the use of IORING_SETUP_SUBMIT_ALL which makes a lot of sense. Unfortunally we never ended up actually enable it as we guarded it with an ifdef that was always false due the usage of an outdated io_uring.h file.

Motifications:

- Check if IORING_SETUP_SUBMIT_ALL is supported by the running system and if this is the case enable it
- Update the io_uring.h file

Result:

IORING_SETUP_SUBMIT_ALL is used when possible