Faster openCursor in NioBuffer (#14689)
Motivation:
When ByteCursor permitted accessing and processing whole longs of data
at a time, and when it was possible to change the byte-order of a Netty
buffer, the byte order of the ByteCursors underlying ByteBuffer
mattered, and as such we had to duplicate the buffer and enforce correct
byte order.

These flexibilities have since been removed, so there's no longer a need
for us to duplicate ByteBuffers for the ByteCursors.

Modification:
Remove ByteBuffer::duplicate calls in NioBuffer.ForwardNioByteCursor and
NioBuffer.ReverseNioByteCursor.

Result:
It's now much faster to open a cursor, and it's more likely for the
cursor object to be scalarized.