ChannelPipeline: If ChannelHandlerContext.isRemoved() returns true we also must return null for ChannelPipeline.context(...)
Motivation:

We might end up in a situation where isRemoved() returns true but the ChannelHandlerContext is still not completely removed from the internal datastructure of DefaultChannelPipeline. In this case we need to ensure we skip it when scanning the handlers in the pipeline to make methods like context(...) and get(...) behave consistent.

Modifications:

Check if the context was removed before consider it as "still in the pipeline".

Result:

Fixes https://github.com/netty/netty/issues/14679