Notify event loop termination future of unexpected exceptions
Motivations:
These exceptions generally shouldn't happen, but if they do, they break internal invariants and the best a system can do is to somehow restart.

To facilitate systems reacting to this, we should tell the termination future about these exceptions.

Modification:
When an event loop terminates from an unexpected exception, we now shut down the event loop as normal, but mark the termination future as failed.

Result:
It's now possible for applications to listen for event loops displaying improper lifetime behavior.

This helps situations such as the one described in https://github.com/netty/netty/issues/14632