Add test to show IoExecutor in action
Motivation:

In 9fbd6993c0a61f53756236728267f966c2750b31 we added the concept of the IoExecutor to remove tight coupling to an EventLoop with the IoHandler. We should provide a testcase that shows that its actually possible now drive an IoHandler without an EventExecutor implementation for more advanced use-cases.

Modifications:

- Add constructor for DefaultPromise which makes it possible to implement an own IoExecutor without the need to also implement a full Promise.
- Add unit test

Result:

Show that its possible to use IoExecutor without a full-blown EventLoop