IoRegistration: Remove usage of Future from external API and allow for wrapping (#14744)
Motivation:

There is really no need for IoRegistration to use a Future in its API.
This simplifies wrapping and also implementations of IoExecutor.

Modifications:

- Remove IoRegistration.cancelFuture() method
- Change IoRegistration.cancel() to return a boolean
- Add IoRegistration.attachment() which might return implementation
specific attachments of this registration (for example for NIO this
returns the underyling SelectionKey)
- Remove throws Exception from IoRegistration methods
- Remove all sub-types (interfaces) of IoRegistration as these are not
needed anymore.

Result:

Easier to implement types and also allow of wrapping. Also remove some
coupling.