Update the pipeline benchmark to make it more realistic (#14724)
Motivation:
If we are to make meaningful optimisations to the pipeline machinery, we
need a benchmark that is pipeline-focused, but still representative of
realistic interaction patterns.

Modification:
Generate a variety of pipelines with many different handlers that
implement a mix of methods, so some will be skipped while others will be
called.

We then measure both single-method call sequences, and we measure a
sequence of different calls.

This puts pressure on the handler-skipping logic.

Result:
We have a more realistic benchmark for message propagation in channel
handler pipelines.