AdaptivePoolingAllocator: Round chunk sizes up to MIN_CHUNK_SIZE units
Motivation:
As the AdaptivePoolingAllocator adjusts it's preferred chunk size, it may end up causing fragmentation with the system allocator.

Modification:
The chunk size is now rounded up to the nearest whole multiple of the MIN_CHUNK_SIZE.
The MIN_CHUNK_SIZE is itself already an even multiple of many popular small page sizes, like 4k, 16k, and 64k, which makes it easier for the system allocator to manage the memory in terms of whole pages.

Result:
Reduced memory fragmentation tendency of the adaptive allocator.