Reduce pipeline stack depth and improve performance (#14705)
Motivation:
Pipeline calls, such as `ctx.fireChannelRead` and `ctx.write`, currently
go through multiple methods. This increases the stack depth in event
loop threads, which makes it harder to debug (people have more ceremony
mixed in with their code), and hurts performance (JIT inliner budget
gets used up faster).

Furthermore, there are other places in this machinery which can be made
faster.

Modification:

**1.**
In the `AbstractChannelHandlerContext` all inbound methods, e.g.
`fireChannelRead`, have been manually inlined. This means that when a
handler calls `ctx.fireChannelRead`, this method call in turn now
directly call the `channelRead` method of the _next_ handler in the
pipeline. Previously, we had two extra method calls here.

As a consequence of this inlining, we now have to re-compute the target
context when we trampoline tasks onto the event-loop. This is presumably
rare, and worth the cost.
This also means that some code now moves from the executor of the target
context, to the executor of the calling context. This can create
different behaviors from Netty 4.1, if the pipeline has multiple
handlers, is modified by the handlers during the call, and the handlers
use child-executors.

**2.**
A few outbound methods - `read`, `write`, `writeAndFlush` - are likewise
inlined. Their inherent complexity and number of overloads means we
can't realistically get them down to a single method call, but we can
get them down to two. This is still a nice improvement.
The `flush` method is usually not implemented by handlers, so there's no
point in inlining that further.

**3.**
In every such call, after finding the next context, we have to decide if
we can call the handler directly, or need to trampoline onto a different
event loop (due to the executor off-loading feature). This means we have
to inspect the context and either pick its child executor, or load the
channel event loop of the target context, and this latter part (which is
the most common case) is a sequence of dependent loads. Dependent loads
cause cache misses and CPU pipeline stalls, so to deal with this the
`AbstractChannelHandlerContext` now has a `contextExecutor` field, which
caches the result of computing the concrete executor. This means our
executor is only one dependent (on the channel handler context) load
away. The speedup from this is quite noticeable because it's such a
common operation.



Results:
The `DefaultChannelPipelineBenchmark` on my M1 Pro, running JDK 17,
tells an encouraging story:

```
Before:
Benchmark                                         (extraHandlers)  (pipelineArrayLength)   Mode  Cnt        Score        Error  Units
DefaultChannelPipelineBenchmark.propagateEvent                 16                   1024  thrpt   10  6515502.983 ± 253375.107  ops/s
DefaultChannelPipelineBenchmark.propagateVariety               16                   1024  thrpt   10  7628162.835 ±  40168.895  ops/s

After:
Benchmark                                         (extraHandlers)  (pipelineArrayLength)   Mode  Cnt        Score        Error  Units
DefaultChannelPipelineBenchmark.propagateEvent                 16                   1024  thrpt   50  6899710.275 ± 115305.284  ops/s
DefaultChannelPipelineBenchmark.propagateVariety               16                   1024  thrpt   50  7985208.748 ±  31532.793  ops/s
```